backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
python-decouple>=3.8
# GeoDjango uses system GDAL/GEOS (provided by PostGIS installer on Windows)
